updated top level server to pull scenes from config and fixed hub.js
made it so that hub.js queries the server for a scene and when the top level server  checks the current scene versus the scenes in the config it will respond with URLs.  If no scenes are found it responds with a noUrls reply that the hub.js catches and breaks out of script injection with.